Add web adapter.

+++ src/main/java/org/test_create_dancing_events/application/port/inbound/CreateDancingEventInput.java
@@ -1,5 +1,5 @@
 package org.test_create_dancing_events.application.port.inbound;
 
-public record CreateDancingEventInput(String title, String description, String eventDate, long eventOrganizerId) {
+public record CreateDancingEventInput(String title, String description, String eventDate, int eventOrganizerId) {
 
 }
